refactor: form state
